begin decimal

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: alafranc <alafranc@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/10 13:29:22 by alafranc          #+#    #+#             */
-/*   Updated: 2020/12/13 02:30:39 by alafranc         ###   ########lyon.fr   */
+/*   Updated: 2020/12/13 18:46:38 by alafranc         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ int				ft_putchar_1(char c);
 int				ft_putnbr_hexa_long_count(long long nb);
 int				ft_putnbr_hexa_count(int nb);
 int				ft_putstr_count(char *str, int size);
-int				ft_putnbr_count(int nb);
+int				ft_putnbr_count(long nb);
 const char		*parse_flags_lmc(const char *format, t_flags *flags, va_list ap);
 const char		*parse_flags_precision(const char *format,
 				t_flags *flags, va_list ap);
